+++ src/components/modals/BinListModal/MobileBinList.tsx
@@ -2,7 +2,7 @@ import { useState, useCallback, useEffect, useRef, useId } from 'react';
 import { createPortal } from 'react-dom';
 import { DEFAULT_CATEGORY_COLOR } from '../../../constants';
 import { useBinList } from '../../../hooks/useBinList';
-import { SplitPreview } from '../../PrintList';
+import { SplitPreview } from '../../Print';
 import type { EnhancedPrintRow, Category, PrintListSortKey } from '../../../types';
 
 interface MobileBinListProps {
