+++ src/components/modals/PrintModal.tsx
@@ -3,8 +3,8 @@ import { createPortal } from 'react-dom';
 import { useShallow } from 'zustand/shallow';
 import { useLayoutStore } from '../../store';
 import { useSettingsStore, type PrintViewSettings, type BinListSortOrder } from '../../store/settings';
-import { PrintLayout } from '../print/PrintLayout';
-import { SortOrderConfig } from '../print/SortOrderConfig';
+import { PrintLayout } from '../Print/PrintLayout';
+import { SortOrderConfig } from '../Print/SortOrderConfig';
 import { Checkbox } from '../Checkbox';
 import { getBinCountByLayer } from '../../utils/printLayout';
 import '../../styles/print.css';
